pyproject.toml
[tool.poetry]
name = "CARTO"
version = "0.1.0"
description = "CARTO's backend challenge"
authors = ["Eugenio <eugeniodonaque@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "2.1.1"
pytest = "^7.1.1"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.0"
PyYAML = "^6.0"
python-dateutil = "^2.8.2"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
